requirements.txt
flask
pillow==11.3.0
transformers==4.56.2
onnxruntime==1.23.0
gunicorn
